internal/assert: new package
Introduce an assertion test helper package to DRY up the many calls to
cmp.Diff.
