+++ internal/e2e/lds_test.go
@@ -26,6 +26,7 @@ import (
 	envoy_api_v2_listener "github.com/envoyproxy/go-control-plane/envoy/api/v2/listener"
 	envoy_config_v2_tcpproxy "github.com/envoyproxy/go-control-plane/envoy/config/filter/network/tcp_proxy/v2"
 	"github.com/envoyproxy/go-control-plane/pkg/wellknown"
+	"github.com/projectcontour/contour/internal/assert"
 	"github.com/projectcontour/contour/internal/contour"
 	"github.com/projectcontour/contour/internal/dag"
 	"github.com/projectcontour/contour/internal/envoy"
@@ -43,7 +44,7 @@ func TestNonTLSListener(t *testing.T) {
 
 	// assert that without any ingress objects registered
 	// there are no active listeners
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "0",
 		Resources: resources(t,
 			staticListener(),
@@ -79,7 +80,7 @@ func TestNonTLSListener(t *testing.T) {
 
 	// add it and assert that we now have a ingress_http listener
 	rh.OnAdd(i1)
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "1",
 		Resources: resources(t,
 			&v2.Listener{
@@ -109,7 +110,7 @@ func TestNonTLSListener(t *testing.T) {
 
 	// update i1 to i2 and verify that ingress_http has gone.
 	rh.OnUpdate(i1, i2)
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "2",
 		Resources: resources(t,
 			staticListener(),
@@ -135,7 +136,7 @@ func TestNonTLSListener(t *testing.T) {
 
 	// update i2 to i3 and check that ingress_http has returned
 	rh.OnUpdate(i2, i3)
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "3",
 		Resources: resources(t,
 			&v2.Listener{
@@ -209,7 +210,7 @@ func TestTLSListener(t *testing.T) {
 	rh.OnAdd(s1)
 
 	// assert that there is only a static listener
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "0",
 		Resources: resources(t,
 			staticListener(),
@@ -220,7 +221,7 @@ func TestTLSListener(t *testing.T) {
 
 	// add ingress and assert the existence of ingress_http and ingres_https
 	rh.OnAdd(i1)
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "1",
 		Resources: resources(t,
 			&v2.Listener{
@@ -271,7 +272,7 @@ func TestTLSListener(t *testing.T) {
 
 	// update i1 to i2 and verify that ingress_http has gone.
 	rh.OnUpdate(i1, i2)
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "2",
 		Resources: resources(t,
 			&v2.Listener{
@@ -290,7 +291,7 @@ func TestTLSListener(t *testing.T) {
 
 	// delete secret and assert that ingress_https is removed
 	rh.OnDelete(s1)
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "3",
 		Resources: resources(t,
 			staticListener(),
@@ -383,7 +384,7 @@ func TestIngressRouteTLSListener(t *testing.T) {
 	rh.OnAdd(secret1)
 
 	// assert that there is only a static listener
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "0",
 		Resources: resources(t,
 			staticListener(),
@@ -409,7 +410,7 @@ func TestIngressRouteTLSListener(t *testing.T) {
 	// add ingress and assert the existence of ingress_http and ingres_https
 	rh.OnAdd(i1)
 
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "1",
 		Resources: resources(t,
 			&v2.Listener{
@@ -429,7 +430,7 @@ func TestIngressRouteTLSListener(t *testing.T) {
 	// delete secret and assert both listeners are removed because the
 	// ingressroute is no longer valid.
 	rh.OnDelete(secret1)
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "2",
 		Resources: resources(t,
 			staticListener(),
@@ -454,7 +455,7 @@ func TestIngressRouteTLSListener(t *testing.T) {
 
 	// add ingress and assert the existence of ingress_http and ingres_https
 	rh.OnAdd(i2)
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "4",
 		Resources: resources(t,
 			&v2.Listener{
@@ -532,7 +533,7 @@ func TestLDSFilter(t *testing.T) {
 
 	// add ingress and fetch ingress_https
 	rh.OnAdd(i1)
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "1",
 		Resources: resources(t,
 			&v2.Listener{
@@ -549,7 +550,7 @@ func TestLDSFilter(t *testing.T) {
 	}, streamLDS(t, cc, "ingress_https"))
 
 	// fetch ingress_http
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "1",
 		Resources: resources(t,
 			&v2.Listener{
@@ -565,7 +566,7 @@ func TestLDSFilter(t *testing.T) {
 	}, streamLDS(t, cc, "ingress_http"))
 
 	// fetch something non existent.
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "1",
 		TypeUrl:     listenerType,
 		Nonce:       "1",
@@ -577,7 +578,7 @@ func TestLDSStreamEmpty(t *testing.T) {
 	defer done()
 
 	// assert that streaming LDS with no ingresses does not stall.
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "0",
 		TypeUrl:     listenerType,
 		Nonce:       "0",
@@ -644,7 +645,7 @@ func TestLDSTLSMinimumProtocolVersion(t *testing.T) {
 
 	// add ingress and fetch ingress_https
 	rh.OnAdd(i1)
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "2",
 		Resources: resources(t,
 			&v2.Listener{
@@ -700,7 +701,7 @@ func TestLDSTLSMinimumProtocolVersion(t *testing.T) {
 	// easier to patch this up than add more params to filterchaintls
 	l1.FilterChains[0].TlsContext.CommonTlsContext.TlsParams.TlsMinimumProtocolVersion = envoy_api_v2_auth.TlsParameters_TLSv1_3
 
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "3",
 		Resources: resources(t,
 			l1,
@@ -718,7 +719,7 @@ func TestLDSIngressHTTPUseProxyProtocol(t *testing.T) {
 
 	// assert that without any ingress objects registered
 	// there is only a static listener
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "0",
 		Resources: resources(t,
 			staticListener(),
@@ -754,7 +755,7 @@ func TestLDSIngressHTTPUseProxyProtocol(t *testing.T) {
 	// add it and assert that we now have a ingress_http listener using
 	// the proxy protocol (the true param to filterchain)
 	rh.OnAdd(i1)
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "1",
 		Resources: resources(t,
 			&v2.Listener{
@@ -819,7 +820,7 @@ func TestLDSIngressHTTPSUseProxyProtocol(t *testing.T) {
 	rh.OnAdd(s1)
 
 	// assert that there is only a static listener
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "0",
 		Resources: resources(t,
 			staticListener(),
@@ -855,7 +856,7 @@ func TestLDSIngressHTTPSUseProxyProtocol(t *testing.T) {
 		),
 		FilterChains: filterchaintls("kuard.example.com", s1, envoy.HTTPConnectionManager("ingress_https", envoy.FileAccessLogEnvoy("/dev/stdout")), "h2", "http/1.1"),
 	}
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "1",
 		Resources: resources(t,
 			&v2.Listener{
@@ -924,7 +925,7 @@ func TestLDSCustomAddressAndPort(t *testing.T) {
 	rh.OnAdd(s1)
 
 	// assert that there is only a static listener
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "0",
 		Resources: resources(t,
 			staticListener(),
@@ -966,7 +967,7 @@ func TestLDSCustomAddressAndPort(t *testing.T) {
 		),
 		FilterChains: filterchaintls("kuard.example.com", s1, envoy.HTTPConnectionManager("ingress_https", envoy.FileAccessLogEnvoy("/dev/stdout")), "h2", "http/1.1"),
 	}
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "1",
 		Resources: resources(t,
 			ingress_http,
@@ -1040,7 +1041,7 @@ func TestLDSCustomAccessLogPaths(t *testing.T) {
 	rh.OnAdd(s1)
 
 	// assert that there is only a static listener
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "0",
 		Resources: resources(t,
 			staticListener(),
@@ -1066,7 +1067,7 @@ func TestLDSCustomAccessLogPaths(t *testing.T) {
 		),
 		FilterChains: filterchaintls("kuard.example.com", s1, envoy.HTTPConnectionManager("ingress_https", envoy.FileAccessLogEnvoy("/tmp/https_access.log")), "h2", "http/1.1"),
 	}
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "1",
 		Resources: resources(t,
 			ingress_http,
@@ -1085,7 +1086,7 @@ func TestLDSIngressRouteInsideRootNamespaces(t *testing.T) {
 	defer done()
 
 	// assert that there is only a static listener
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "0",
 		Resources: resources(t,
 			staticListener(),
@@ -1131,7 +1132,7 @@ func TestLDSIngressRouteInsideRootNamespaces(t *testing.T) {
 	rh.OnAdd(ir1)
 
 	// assert there is an active listener
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "1",
 		Resources: resources(t,
 			&v2.Listener{
@@ -1155,7 +1156,7 @@ func TestLDSIngressRouteOutsideRootNamespaces(t *testing.T) {
 	defer done()
 
 	// assert that there is only a static listener
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "0",
 		Resources: resources(t,
 			staticListener(),
@@ -1186,7 +1187,7 @@ func TestLDSIngressRouteOutsideRootNamespaces(t *testing.T) {
 	rh.OnAdd(ir1)
 
 	// assert that there is only a static listener
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "1",
 		Resources: resources(t,
 			staticListener(),
@@ -1201,7 +1202,7 @@ func TestIngressRouteHTTPS(t *testing.T) {
 	defer done()
 
 	// assert that there is only a static listener
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "0",
 		Resources: resources(t,
 			staticListener(),
@@ -1285,7 +1286,7 @@ func TestIngressRouteHTTPS(t *testing.T) {
 		),
 		FilterChains: filterchaintls("example.com", s1, envoy.HTTPConnectionManager("ingress_https", envoy.FileAccessLogEnvoy("/dev/stdout")), "h2", "http/1.1"),
 	}
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "1",
 		Resources: resources(t,
 			ingressHTTP,
@@ -1355,7 +1356,7 @@ func TestLDSIngressRouteTCPProxyTLSPassthrough(t *testing.T) {
 		),
 	}
 
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "1",
 		Resources: resources(t,
 			ingressHTTPS,
@@ -1428,7 +1429,7 @@ func TestLDSIngressRouteTCPForward(t *testing.T) {
 		),
 	}
 
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "1",
 		Resources: resources(t,
 			ingressHTTPS,
@@ -1445,7 +1446,7 @@ func TestIngressRouteTLSCertificateDelegation(t *testing.T) {
 	defer done()
 
 	// assert that there is only a static listener
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "0",
 		Resources: resources(t,
 			staticListener(),
@@ -1507,7 +1508,7 @@ func TestIngressRouteTLSCertificateDelegation(t *testing.T) {
 	})
 
 	// assert there are no listeners
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "1",
 		Resources: resources(t,
 			staticListener(),
@@ -1550,7 +1551,7 @@ func TestIngressRouteTLSCertificateDelegation(t *testing.T) {
 		FilterChains: filterchaintls("example.com", s1, envoy.HTTPConnectionManager("ingress_https", envoy.FileAccessLogEnvoy("/dev/stdout")), "h2", "http/1.1"),
 	}
 
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "2",
 		Resources: resources(t,
 			ingress_http,
@@ -1578,7 +1579,7 @@ func TestIngressRouteTLSCertificateDelegation(t *testing.T) {
 	}
 	rh.OnUpdate(t1, t2)
 
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "3",
 		Resources: resources(t,
 			ingress_http,
@@ -1606,7 +1607,7 @@ func TestIngressRouteTLSCertificateDelegation(t *testing.T) {
 	}
 	rh.OnUpdate(t2, t3)
 
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "4",
 		Resources: resources(t,
 			staticListener(),
@@ -1632,7 +1633,7 @@ func TestIngressRouteTLSCertificateDelegation(t *testing.T) {
 	}
 	rh.OnUpdate(t3, t4)
 
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "5",
 		Resources: resources(t,
 			staticListener(),
@@ -1715,7 +1716,7 @@ func TestIngressRouteMinimumTLSVersion(t *testing.T) {
 	l1.FilterChains[0].TlsContext.CommonTlsContext.TlsParams.TlsMinimumProtocolVersion = envoy_api_v2_auth.TlsParameters_TLSv1_2
 
 	// verify that i1's TLS 1.1 minimum has been upgraded to 1.2
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "1",
 		Resources: resources(t,
 			&v2.Listener{
@@ -1768,7 +1769,7 @@ func TestIngressRouteMinimumTLSVersion(t *testing.T) {
 	l2.FilterChains[0].TlsContext.CommonTlsContext.TlsParams.TlsMinimumProtocolVersion = envoy_api_v2_auth.TlsParameters_TLSv1_3
 
 	// verify that i2's TLS 1.3 minimum has NOT been downgraded to 1.2
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "2",
 		Resources: resources(t,
 			&v2.Listener{
@@ -1847,7 +1848,7 @@ func TestLDSIngressRouteRootCannotDelegateToAnotherRoot(t *testing.T) {
 
 	// verify that port 80 is present because while it is not possible to
 	// delegate to it, child can host a vhost which opens port 80.
-	assertEqual(t, &v2.DiscoveryResponse{
+	assert.Equal(t, &v2.DiscoveryResponse{
 		VersionInfo: "2",
 		Resources: resources(t,
 			&v2.Listener{
